Revert "MQTT: Use float for HEATPUMP_POWER"
This reverts commit 5a923c9e69086d2f2cf0826f51f5d2b71cdacefc.

+++ src/services/mqttMapper.ts
@@ -175,7 +175,7 @@ export class MqttMapper {
       return {
         topic: this.answers.mqtt_heatpump_power,
         ...this.sensorBuilder.splittedSensor('HEATPUMP_POWER'),
-        type: 'float',
+        type: 'integer',
       }
   }
